+++ casual/casual-api/src/main/java/se/laz/casual/jca/RuntimeInformation.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023, The casual project. All rights reserved.
+ * Copyright (c) 2023 - 2025, The casual project. All rights reserved.
  *
  * This software is licensed under the MIT license, https://opensource.org/licenses/MIT
  */
@@ -13,6 +13,7 @@ public class RuntimeInformation
 {
     private static final String INBOUND_SERVER_STARTED = "INBOUND_SERVER_STARTED";
     private static final String EVENT_SERVER_STARTED = "EVENT_SERVER_STARTED";
+    private static final String DOMAIN_IS_BEING_SHUTDOWN = "DOMAIN_IS_BEING_SHUTDOWN";
     private static final Map<String, Boolean> CACHE = new ConcurrentHashMap<>();
 
     private RuntimeInformation()
@@ -36,5 +37,15 @@ public class RuntimeInformation
     public static void setEventServerStarted(boolean started)
     {
         CACHE.put(EVENT_SERVER_STARTED, started);
+    }
+
+    public static void setDomainIsBeingShutdown(boolean beingShutdown)
+    {
+        CACHE.put(DOMAIN_IS_BEING_SHUTDOWN, beingShutdown);
+    }
+
+    public static boolean isDomainBeingShutdown()
+    {
+        return Optional.ofNullable(CACHE.get(DOMAIN_IS_BEING_SHUTDOWN)).orElse(false);
     }
 }