+++ casual/casual-jca/src/main/java/se/laz/casual/jca/CasualResourceAdapter.java
@@ -26,6 +26,7 @@ import se.laz.casual.config.ReverseInbound;
 import se.laz.casual.event.server.EventServer;
 import se.laz.casual.event.server.EventServerConnectionInformation;
 import se.laz.casual.jca.inflow.CasualActivationSpec;
+import se.laz.casual.jca.inflow.CasualInboundTransactionRegistry;
 import se.laz.casual.jca.jmx.JMXStartup;
 import se.laz.casual.jca.work.StartInboundServerListener;
 import se.laz.casual.jca.work.StartInboundServerWork;
@@ -68,6 +69,7 @@ public class CasualResourceAdapter implements ResourceAdapter, ReverseInboundLis
     private static Logger log = Logger.getLogger(CasualResourceAdapter.class.getName());
     private ConcurrentHashMap<Integer, CasualActivationSpec> activations = new ConcurrentHashMap<>();
     private List<ReverseInboundServer> reverseInbounds = new ArrayList<>();
+    private CasualInboundTransactionRegistry inboundTransactionRegistry;
     // it is not really unused, it should never ever be gc:ed, thus it is part of this class
     @SuppressWarnings("java:S1068")
     private EventServer eventServer;
@@ -133,6 +135,8 @@ public class CasualResourceAdapter implements ResourceAdapter, ReverseInboundLis
                                    ActivationSpec spec) throws ResourceException
     {
         log.info(()->"start endpointActivation() ");
+        inboundTransactionRegistry = new CasualInboundTransactionRegistry();
+        RuntimeInformation.setDomainIsBeingShutdown(false);
         CasualActivationSpec as = (CasualActivationSpec) spec;
         as.setPort( getInboundServerPort() );
         ConnectionInformation ci = ConnectionInformation.createBuilder()
@@ -141,6 +145,7 @@ public class CasualResourceAdapter implements ResourceAdapter, ReverseInboundLis
                 .withWorkManager(workManager)
                 .withXaTerminator(xaTerminator)
                 .withUseEpoll( ConfigurationService.getConfiguration( ConfigurationOptions.CASUAL_INBOUND_USE_EPOLL ) )
+                .withInboundTransactionRegistry(inboundTransactionRegistry)
                 .build();
         activations.put(as.getPort(), as);
         log.info(() -> "start casual inbound server" );
@@ -164,6 +169,7 @@ public class CasualResourceAdapter implements ResourceAdapter, ReverseInboundLis
                                                                    .withXaTerminator(xaTerminator)
                                                                    .withProtocolVersion(ProtocolVersion.VERSION_1_0)
                                                                    .withMaxBackoffMillils(instance.getMaxConnectionBackoffMillis())
+                                                                   .withInboundTransactionRegistry(inboundTransactionRegistry)
                                                                    .build(), instance.getSize());
         }
     }
@@ -219,10 +225,15 @@ public class CasualResourceAdapter implements ResourceAdapter, ReverseInboundLis
     public void endpointDeactivation(MessageEndpointFactory endpointFactory,
                                      ActivationSpec spec)
     {
-        log.finest(()->"endpointDeactivation()");
+        log.info(()->"endpointDeactivation() ");
+        RuntimeInformation.setDomainIsBeingShutdown(true);
         InboundDeactivatedContext.domainDisconnect();
         InboundDeactivatedContext.clear();
         InboundTopologyUpdateContext.clear();
+        Predicate predicate = () -> inboundTransactionRegistry.hasPending() || CasualResourceManager.getInstance().hasPending();
+        long sleepTimeMilliseconds = 20L;
+        ShutdownBarrier shutdownBarrier = ShutdownBarrier.of(sleepTimeMilliseconds, predicate);
+        shutdownBarrier.intermittentSleep();
         if( server != null )
         {
             server.close();