+++ casual/casual-jca/src/main/java/se/laz/casual/jca/CasualResourceManager.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017 - 2024, The casual project. All rights reserved.
+ * Copyright (c) 2017 - 2025, The casual project. All rights reserved.
  *
  * This software is licensed under the MIT license, https://opensource.org/licenses/MIT
  */
@@ -42,6 +42,11 @@ public final class CasualResourceManager
             throw new CasualResourceAdapterException("xid: " + xid + " already stored for domain: " + domainId);
         }
         pendingRequests.computeIfAbsent(domainId, k -> ConcurrentHashMap.newKeySet()).add(xid);
+    }
+
+    public synchronized boolean hasPending()
+    {
+        return !pendingRequests.isEmpty();
     }
 
     public synchronized void remove(DomainId domainId, final Xid xid)