+++ casual/casual-network/src/test/groovy/se/laz/casual/network/inbound/ExceptionHandlerTest.groovy
@@ -9,6 +9,7 @@ package se.laz.casual.network.inbound
 import io.netty.channel.Channel
 import io.netty.channel.ChannelHandlerContext
 import io.netty.channel.ChannelId
+import se.laz.casual.jca.inflow.CasualInboundTransactionRegistry
 import spock.lang.Specification
 
 class ExceptionHandlerTest extends Specification
@@ -16,12 +17,13 @@ class ExceptionHandlerTest extends Specification
     def 'should close context'()
     {
         setup:
+        CasualInboundTransactionRegistry inboundTransactionRegistry = new CasualInboundTransactionRegistry()
         def ctx = Mock(ChannelHandlerContext){
            channel() >> Mock(Channel){
               id() >> Mock(ChannelId)
            }
         }
-        def handler = ExceptionHandler.of()
+        def handler = ExceptionHandler.of(inboundTransactionRegistry)
         when:
         handler.exceptionCaught(ctx, new RuntimeException())
         then: